src/main/java/com/kakaopay/bank/bank/security/JwtTokenChecker.java
package com.kakaopay.bank.bank.security;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jws;
import io.jsonwebtoken.Jwts;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

import java.io.UnsupportedEncodingException;

@Component
public class JwtTokenChecker implements TokenChecker {
    private @Value("${jwt.secret}")
    String secretKey;

    @Override
    public boolean isUsable(String jwt) {
        try{
            Jws<Claims> claims = Jwts.parser()
                                     .setSigningKey(this.generateKey())
                                     .parseClaimsJws(jwt);
            return true;
        }catch (Exception e) {
            throw new UnauthorizedException(e.getMessage());
        }
    }

    private byte[] generateKey() {
        byte[] key = null;
        try {
            key = secretKey.getBytes("UTF-8");
        } catch (UnsupportedEncodingException e) {
            e.printStackTrace();
        }
        return key;
    }

}
